+++ project/datamodules/text/text_classification_test.py
@@ -5,7 +5,7 @@ import pytest
 
 from project.datamodules.text.text_classification import TextClassificationDataModule
 from project.experiment import instantiate_datamodule
-from project.utils.hydra_config_utils import get_config_loader
+from project.utils.testutils import get_config_loader
 
 datamodule_configs = ["glue_cola"]
